Change/Fix: Paths to groups
Had to bite the bullet and insert the element paths individually in
the structs.  It's the most clear and reliable way to save and access
them.  Feels messy, but it works.

AFAIK /everything/ works now, whether a frame's workspace is a
top-level group or a lower-level one, and whether there are nil heads
in the path or not.

Maybe this can be simplified, or factored out into bufler-group-tree.el
sometime, but let's not break anything for a while.