Fix: (bufler-workspace-tabs)
Just fix it.  I think I have spent more time on this function than
anything else in the whole package by now.  All because of the
decision to allow paths to start with nil so subgroups can be raised
up a level.  It's a good feature, but I did not foresee how
complicated it would make the implementation.  In the future I will
try to use a struct for groups, with paths and elements as slots,
which I hope will make all this much easier.

Fix #9.  Thanks to @adouzzy for reporting.