Replace pretty-hydra with transient
Since this whole package is already based on magit-section, it makes
sense to keep the user interface consistent and use transient (a library
created for and used extensively by the Magit project).

I would like to see the transient definition just be auto-created from
the keymap, or vice versa. As far as I can tell, this is only defined in
order to provide a pretty "help" menu. Still, I don't use any other
package which pulls in Hydra, so I'd like to avoid that dependency.