Fix: (helm-bufler-source) Remove initial-nil-skipping group path
I don't remember why this was necessary, but it seems to be incorrect
now: with it, certain bufler-groups definitions (like the default one)
don't work properly (even though my personal config does), but without
it, they do (both the default and my personal config).  So removing it
seems like the right thing to do.