Add: (bufler-project-current) Cache project metadata
When many files were open, I noticed that bufler's grouping could seem
slow.  Profiling revealed that the `project-current' call-tree (used
in the `project' and `parent-project' auto groups) caused a lot of
consing and GC, apparently due to its calling `directory-files' (which
is noticeable in large directories).  Memoizing `project-current'
solves this.

Unfortunately the new `with-memoization' macro in Emacs 29 isn't
suitable since it can't distinguish between nil values and not-found
keys in hash tables.